requirements.txt
# Core dependencies
slack-sdk>=3.36.0
PyYAML>=6.0.2
asyncio-mqtt>=0.11.0

# Development dependencies
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
pytest-cov>=5.0.0

# Optional dependencies for enhanced functionality
croniter>=2.0.2
psutil>=5.9.8
aiofiles>=23.2.1
aiohttp>=3.9.0

# Type hinting
mypy>=1.10.0
types-PyYAML
types-croniter
types-psutil
